Refactor mapping system to adapt for recent TestSwarm changes.
The browserSets system has changed in jquery/testswarm. This was
done to make it possible to make use of OS and devices.

As a side-effect it is now possible to associate a worker browser
template to a testswarm uaID without needing to manually map them.

I've therefore dropped most of the mapping in favour of automatic
detection. This should make maintenance a lot easier and allow
continued use of TestSwarm and testswarm-browserstack without
needing any changes in the software (all you need is to update
your local submission settings to include the browsers you want
to and make sure that you are able to provide those browsers,
which in this case means BrowserStack supports them).

* No more registering in jquery/testswarm useragents.ini.
* No more updating map.js in testswarm-browserstack.

Related PR in TestSwarm that this is related to:
https://github.com/jquery/testswarm/pull/244