Logging: Improve error logging.
Log all errors with 'action=error' and a non-variable 'message'
property that can be searched for (all variables are in the 'info' property).

Made sure all util.log calls now use the object format.
There were a few that were logging raw messages, however only
in methods that were only called from cli.js. So I moved these to
there to avoid accidental mixup.

Fixed bug where --run-loop output had spurious new lines between
messages. This was caused by runLoop() using console methods on
data that was already formatted by console methods in the child
process. Part of the formatting is ending it in a line break, so
this happened twice.

This will make it much easier to spot warnings and fatals in
the logs.