Add points for missing osMinor & osPatch when missing also on BrowserStack
Handle the "os" field in a special way - some browsers on BrowserStack only
define the major part and we want to be able to prefer them when there's
another entry with a matching major that also specifies the minor.
One example is iOS: BrowserStack includes both "10" & "10.3"; the former
is a real device and the latter is an emulator; we want the real device.