Skip browsers with non-numerical versions like Edge Insider Preview
BrowserStack has just started sharing Edge Insider Preview via its API. That
browser has `browser_version` set to "insider preview". This commit makes us
skip browsers with non-numerical versions like that. Currently, only Edge
Insider Preview violates this constraint and there's hardly any reason to
test in it - this is a preview of the legacy version of Edge which is not
receiving updates anymore.

This change is needed to stop crashing testswarm-browserstack completely when
the API contains a browser with such a non-conforming version.